Reorganize `drawing` module
