--- snapper/src/drawing/drawable.rs
@@ -1,12 +1,0 @@
-use tiny_skia::Pixmap;
-
-use crate::Snapper;
-
-pub trait Drawable {
-    fn draw(
-        &self,
-        snapper: &Snapper,
-        pixmap: &mut Pixmap,
-        center: geo::Point,
-    ) -> Result<(), crate::Error>;
-}
