+++ snapper/src/drawing/style/mod.rs
@@ -1,0 +1,22 @@
+use tiny_skia::Color;
+
+pub mod geo;
+
+#[derive(Clone, Debug, PartialEq)]
+pub struct ColorOptions {
+    pub foreground: Color,
+    pub background: Color,
+    pub anti_alias: bool,
+    pub border: Option<f32>,
+}
+
+impl Default for ColorOptions {
+    fn default() -> Self {
+        Self {
+            foreground: Color::from_rgba8(248, 248, 248, 255),
+            background: Color::from_rgba8(26, 26, 26, 255),
+            anti_alias: true,
+            border: Some(1.0),
+        }
+    }
+}
